requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
google-genai
pinecone-client
faiss-cpu
pandas
numpy
pypdf2
python-docx
requests
sentence-transformers
langchain
langchain-google-genai
langchain-pinecone
fastapi_cache
nltk